Support for resuming interrupted experiments
Especially for the case when Metacentrum kills the task after time limit

Only needed thing is to add a field

  'resume',             { true }

into the experiment definition into the first 'bbobParams' section (and
'expInit' the experiment). After that, the bbob_test_01.m will first
look into the $EXPPATH_SHORT/$EXPID/ for the $EXP_ID_tmp_$ID.mat file
and if it exists, it will load it and continue with unfinished
instances.