Ranking Error Measure (inspired by Kern, 2006)
Error measure of a ranking vector. The error is calculated
as sum of the differences between its position in the vector
(provided rank) and the rank which corresponds to its
value (true rank).

errRankMu.m --  the ordering among elements (mu+1),(mu+2)...(length(x))
                is not considered as being error and positioning them
                among the first mu positions is calculated for only one
                error-point; normalized to <0,1>
errRankMuOnly.m -- rank-errors are calculated only for the first 'mu'
                ranked values (the same as in Kern, 2006)
expectedRankDiff.m -- returns ordering of points sorted in the order of
                expected ranking error (calculated by errRankMuOnly())