DTS-CMA-ES RankDiff adaptation
Adaptation of the number of original-evaluated points based on
Ranking difference error between the first and the second model in the
DTS generation.

RankDiff error is transformed to the original-evaluated ratio (of
lambda) based on a linear transfer function which is specified by four
parameters lowRank, highRank, minRatio and maxRatio. This 'ratio' is then
exponentially smothed via
  r_{t} = (1 - a)*r_{t-1} + a*ratio
and
  ceil(r_{t} * lambda)
points are taken for reevaluation. The new points are re-evaluated until
the number of re-evaluted points is lower or equal to the calculated
number of points based on the RankDiff error.

   exp/log/DTFileStatistics.m   -- minor change: new statistics
   exp/log/DTScreenStatistics.m -- minor change: new statistics
   exp/opt_cmaes.m              -- standard population size in pure CMA-ES experiments
   src/surrogate/DoubleTrainedEC.m      -- new update strategy, cycle of re-evaluations
   src/surrogate/OrigRatioUpdaterFactory.m      -- RankDiff-error-based Updater
   src/surrogate/OrigRatioUpdaterRankDiff.m     -- RankDiff-error-based Updater
   src/surrogate/Population.m   -- initial phase := (-1)
   src/util/errRankMu.m         -- errRankMuOnly directly for two y-values vectors